Create util script
