+++ server/feature_hover.py
@@ -16,6 +16,7 @@ import re
 from typing import Optional
 from spacy import registry
 from .spacy_server import SpacyLanguageServer
+from .util import get_current_word
 
 
 def hover(
@@ -76,37 +77,6 @@ def hover(
 
     except Exception as e:
         return None
-
-
-def get_current_word(line: str, start_pos: int):
-    """
-    Returns the a span string seperated by non-word characters
-
-    EXAMPLE OUTPUT:
-    ('architectures', 1, 13)
-    ('spacy', 18, 22)
-    """
-
-    # Verify index lies within boundaries
-    if start_pos < 0 or start_pos >= len(line):
-        return "", 0, 0
-
-    end_i = start_pos
-    start_i = start_pos
-
-    for i in range(start_pos, len(line), 1):
-        if re.match("\W", line[i]):
-            break
-        else:
-            end_i = i
-
-    for i in range(start_pos, -1, -1):
-        if re.match("\W", line[i]):
-            break
-        else:
-            start_i = i
-
-    return line[start_i : end_i + 1], start_i, end_i
 
 
 def detect_registry_names(line: str, word_start: int, word_end: int):
